backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.0
supabase==2.5.1
gunicorn==22.0.0
